requirement.txt
Werkzeug==2.3.6
alembic==1.12.0
attrs==23.1.0
blinker==1.6.3
cachelib==0.10.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
cloudpickle==2.2.1
contourpy==1.1.1
cycler==0.12.1
dacite==1.8.1
databricks-cli==0.18.0
docker==6.1.3
entrypoints==0.4
Flask==2.3.2
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-Session==0.5.0
Flask-SQLAlchemy==3.1.1
fonttools==4.43.1
gitdb==4.0.10
GitPython==3.1.40
gunicorn==21.2.0
htmlmin==0.1.12
idna==3.4
ImageHash==4.3.1
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
llvmlite==0.41.1
Mako==1.2.4
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.6.3
mlflow==2.7.1
multimethod==1.9.1
networkx==3.2
numba==0.58.1
numpy==1.23.5
oauthlib==3.2.2
packaging==23.2
pandas==1.5.3
pandas-profiling==3.6.6
patsy==0.5.3
phik==0.12.3
Pillow==10.1.0
plotly==5.17.0
protobuf==4.24.4
pyarrow==13.0.0
pydantic==1.10.13
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
querystring-parser==1.2.4
requests==2.28.2
scikit-learn==1.3.1
scipy==1.9.3
seaborn==0.12.2
shap==0.43.0
six==1.16.0
slicer==0.0.7
smmap==5.0.1
SQLAlchemy==2.0.22
sqlparse==0.4.4
statsmodels==0.13.5
tabulate==0.9.0
tangled-up-in-unicode==0.2.0
tenacity==8.2.3
threadpoolctl==3.2.0
tqdm==4.64.1
typeguard==2.13.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
visions==0.7.5
websocket-client==1.6.4
wordcloud==1.9.2
ydata-profiling==4.1.2
zipp==3.17.0